Moved customised menus to own package.
