+++ src/main/java/com/kuborros/FurBotNeo/commands/ShopCommands/BuyRoleCommand.java
@@ -5,8 +5,8 @@ import com.jagrosh.jdautilities.command.CommandEvent;
 import com.jagrosh.jdautilities.commons.waiter.EventWaiter;
 import com.jagrosh.jdautilities.doc.standard.CommandInfo;
 import com.jagrosh.jdautilities.examples.doc.Author;
+import com.kuborros.FurBotNeo.utils.menus.StoreDialog;
 import com.kuborros.FurBotNeo.utils.store.ShopItem;
-import com.kuborros.FurBotNeo.utils.store.StoreDialog;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.MessageBuilder;
 import net.dv8tion.jda.api.Permission;
