+++ src/main/java/com/kuborros/FurBotNeo/commands/ShopCommands/SetRoleCommand.java
@@ -5,7 +5,7 @@ import com.jagrosh.jdautilities.command.CommandEvent;
 import com.jagrosh.jdautilities.commons.waiter.EventWaiter;
 import com.jagrosh.jdautilities.doc.standard.CommandInfo;
 import com.jagrosh.jdautilities.examples.doc.Author;
-import com.kuborros.FurBotNeo.utils.store.SelectionTitleDialog;
+import com.kuborros.FurBotNeo.utils.menus.SelectionTitleDialog;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
